Add Reception and assignement step to investigation cases
