+++ apps/judicial-system/web/messages/InvestigationCases/Court/hearingArrangements.ts
@@ -24,40 +24,6 @@ export const icHearingArrangements = {
     description: 'Notaður sem titill á fyrirtöku skrefi í rannsóknarheimildum.',
   }),
   sections: {
-    setJudge: defineMessages({
-      title: {
-        id:
-          'judicial.system.investigation_cases:hearing_arrangements.set_judge.title',
-        defaultMessage: 'Dómari',
-        description:
-          'Notaður sem titll fyrir "dómari" hlutann á fyrirtöku skrefi í rannsóknarheimildum.',
-      },
-      tooltip: {
-        id:
-          'judicial.system.investigation_cases:hearing_arrangements.set_judge.tooltip',
-        defaultMessage:
-          'Dómarinn sem er valinn hér verður skráður á málið og mun fá tilkynningar sendar í tölvupóst. Eingöngu skráður dómari getur svo undirritað úrskurð.',
-        description:
-          'Notaður sem upplýsingatexti í upplýsingasvæði við "dómari" titlinn á fyrirtöku skrefi í rannsóknarheimildum.',
-      },
-    }),
-    setRegistrar: defineMessages({
-      title: {
-        id:
-          'judicial.system.investigation_cases:hearing_arrangements.set_registrar.title',
-        defaultMessage: 'Dómritari',
-        description:
-          'Notaður sem titll fyrir "Dómritari" hlutann á fyrirtöku skrefi í rannsóknarheimildum.',
-      },
-      tooltip: {
-        id:
-          'judicial.system.investigation_cases:hearing_arrangements.set_registrar.tooltip',
-        defaultMessage:
-          'Dómritari sem er valinn hér verður skráður á málið og mun fá tilkynningar sendar í tölvupósti.',
-        description:
-          'Notaður sem upplýsingatexti í upplýsingasvæði við "dómritari" titlinn á fyrirtöku skrefi í rannsóknarheimildum.',
-      },
-    }),
     sessionArrangements: {
       heading: defineMessage({
         id:
