+++ apps/judicial-system/web/pages/domur/rannsoknarheimild/mottaka/[id].ts
@@ -1,0 +1,3 @@
+import ReceptionAndAssignment from '@island.is/judicial-system-web/src/routes/Court/InvestigationRequest/ReceptionAndAssignment/ReceptionAndAssignment'
+
+export default ReceptionAndAssignment
