+++ apps/judicial-system/web/src/routes/Court/InvestigationRequest/HearingArrangements/HearingArrangements.tsx
@@ -1,14 +1,11 @@
 import React, { useContext, useEffect } from 'react'
-import { useQuery } from '@apollo/client'
 
 import { PageLayout } from '@island.is/judicial-system-web/src/components'
 import { SessionArrangements } from '@island.is/judicial-system/types'
 import {
   CourtSubsections,
   Sections,
-  UserData,
 } from '@island.is/judicial-system-web/src/types'
-import { UsersQuery } from '@island.is/judicial-system-web/src/utils/mutations'
 import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import { UserContext } from '@island.is/judicial-system-web/src/components/UserProvider/UserProvider'
 import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
@@ -26,11 +23,6 @@ const HearingArrangements = () => {
   const { user } = useContext(UserContext)
 
   const { autofill, autofillSessionArrangements } = useCase()
-
-  const { data: users, loading: userLoading } = useQuery<UserData>(UsersQuery, {
-    fetchPolicy: 'no-cache',
-    errorPolicy: 'all',
-  })
 
   useEffect(() => {
     document.title = 'Fyrirtaka - Réttarvörslugátt'
@@ -72,12 +64,10 @@ const HearingArrangements = () => {
       isLoading={isLoadingWorkingCase}
       notFound={caseNotFound}
     >
-      {user && users && (
+      {user && (
         <HearingArrangementsForm
           workingCase={workingCase}
           setWorkingCase={setWorkingCase}
-          isLoading={userLoading}
-          users={users}
           user={user}
         />
       )}
